renamed package momentmanager to datamanagers
